requirements.txt
djangorestframework==3.14.0
Django==4.1.7
psycopg2==2.9.5
Python==3.11.2
